% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{greyscale_filter}
\alias{greyscale_filter}
\title{Greyscale filter}
\usage{
greyscale_filter(rgb_vec)
}
\arguments{
\item{rgb_vec}{A vector of 3 raw values representing Red, Green, and Blue.}
}
\value{
A greyscale RGB vector as a raw vector of length 3.
}
\description{
Applies a greyscale transformation to a single RGB vector.
}
